+++ modules/tinymce/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Deleting content within selections that contained content outside a table and content within a cell could result in the entire contents of the cell being incorrectly deleted #TINY-7596
+
 ## 5.9.1 - 2021-08-27
 
 ### Fixed